解决Bug 35，调用vsnprintf时触发警告4389: “==”: 有符号/无符号不匹配（感谢  zhaooptimus）

+++ ucrt/10.0.10240.0/stdio.h
@@ -1437,7 +1437,7 @@ _ACRTXPINLINE int __CRTDECL vsnprintf(
 	if (Count <= 0)
 		return Count;
 
-	if (Count == _BufferCount)
+	if ((size_t)Count == _BufferCount)
 	{
 		_Buffer[Count - 1] = '\0';
 	}
